Final Project[COMPLETE]
This is the final program for my project. We include a number of features, including added adjective, cognate, false cognate, and lexical diversity analysis. We were asked to do verb tense analysis as well, but we were unable to find a good way to accurately count the tenses.

Authors on this are thezombiebandwagon and @jwumb0